projects/payment-system/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-extensions==3.2.3
